# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
      "description": "Vitto Mini Bank é uma API REST para caixas eletrônicos",
      "version": "1.0.0",
      "title": "Vitto Mini Bank API",
      "contact": {
          "email": "annibalhsouza@gmail.com"
      },
      "license": {
          "name": "Apache 2.0",
          "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
  },
  "schemes": ["http"],
  "host": "localhost:3081",
  "basePath": "/api",
  "paths" : {
      "/clientes" : {
          "get" : {
              "summary" : "Obter clientes Vittor",
              "description": "Pegue todos clientes registrados",
              "produces": ["application/json"],
              "parameters": [],
              "responses": {
                  "200": {
                      "description": "successful operation",
                      "schema": {
                          "type": "array",
                          "items": {
                              "$ref": "#/definitions/todosResponse"
                          }
                      }
                  },
                  "400": {
                      "description": "Invalid status value",
                      "schema": {
                          "$ref": "#/definitions/InvalidResponse"
                      }
                  }
              }
          }
      },
      "/cliente" : {
          "post" : {
              "summary" : "Novo cliente",
              "description": "Salvar um novo cliente",
              "produces": ["application/json"],
              "consumes": ["application/json"],
              "parameters": [
                  {
                      "in": "body",
                      "name": "body",
                      "description": "cliente object",
                      "required": true,
                      "schema": {
                          "type": "object",
                          "properties": {
                              "cliente" : {
                                  "type": "object",
                                  "$ref": "#/definitions/Cliente"
                              } 
                          }
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "successful operation",
                      "schema": {
                          "type": "array",
                          "items": {
                              "$ref": "#/definitions/todosResponse"
                          }
                      }
                  },
                  "400": {
                      "description": "Invalid status value",
                      "schema": {
                          "$ref": "#/definitions/InvalidResponse"
                      }
                  }
              }
          }
      },
      "/cliente/" : {
          "put" : {
              "summary" : "Atualizar cliente",
              "description": "Alterar mudanças no cliente Vitto",
              "produces": ["application/json"],
              "parameters": [
                  {
                      "name": "id",
                      "in": "path",
                      "description": "Id do cliente que precisa ser alterado",
                      "required": true,
                      "type": "string"
                  },
                  {
                      "in": "body",
                      "name": "body",
                      "description": "cliente object",
                      "required": true,
                      "schema": {
                          "type": "object",
                          "properties": {
                              "cliente" : {
                                  "type": "object",
                                  "$ref": "#/definitions/Cliente"
                              } 
                          }
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "successful operation",
                      "schema": {
                          "type": "array",
                          "items": {
                              "$ref": "#/definitions/todosResponse"
                          }
                      }
                  },
                  "400": {
                      "description": "Invalid status value",
                      "schema": {
                          "$ref": "#/definitions/InvalidResponse"
                      }
                  }
              }
          }
      },
      "/cliente/{id}" : {
          "delete" : {
              "summary" : "Apagar Cliente",
              "description": "Excluir o cliente desejado",
              "produces": ["application/json"],
              "parameters": [
                  {
                      "name": "id",
                      "in": "path",
                      "description": "Id do cliente que precisa ser excluído",
                      "required": true,
                      "type": "string"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "successful operation",
                      "schema": {
                          "type": "array",
                          "items": {
                              "$ref": "#/definitions/todosResponse"
                          }
                      }
                  },
                  "400": {
                      "description": "Invalid status value",
                      "schema": {
                          "$ref": "#/definitions/InvalidResponse"
                      }
                  }
              }
          }
      }
  }, 
  "definitions": {
      "todosResponse": {
          "type": "object",
          "properties": {
            "_id": {
             "type": "integer"
            },
            "name": {
              "type": "string"
            },
            "cpf": {
              "type": "number"
            },
            "numero_conta": {
              "type": "number"
            },
            "telefone": {
              "type": "number"
            },
            "email": {
              "type": "string"
            },
            "cep": {
              "type": "number"
            },
            "logradouro": {
              "type": "string"
            },
            "estado": {
              "type": "string"
            },
            "cidade": {
              "type": "string"
            },
            "bairro": {
              "type": "string"
            },
            "createdAt": {
              "type": "string"
            },
            "updatedAt": {
              "type": "string"
            },
            "updatedBy": {
              "type": "string"
            }
          }
      },
      "Cliente": {
          "type": "object",
          "properties": {
              "cliente": {
                  "type": {
                      "name": {
                          "type": "string"
                      },
                      "cpf": {
                        "type": "number"
                      },
                      "numero_conta": {
                        "type": "number"
                      },
                      "telefone": {
                        "type": "number"
                      },
                      "email": {
                        "type": "string"
                      },
                      "cep": {
                        "type": "number"
                      },
                      "logradouro": {
                        "type": "string"
                      },
                      "estado": {
                        "type": "string"
                      },
                      "cidade": {
                        "type": "string"
                      },
                      "bairro": {
                        "type": "string"
                      },
                      "createDate": {
                        "type": "string"
                      },
                      "updatedDate": {
                        "type": "string"
                      },
                      "updatedBy": {
                        "type": "string"
                      }
                  }
              }
          }
      },
      "InvalidResponse": {
          "type": "object",
          "properties": {
              "statusCode": {
                  "type": "string"
              },
              "message": {
                  "type": "string"
              }
          }

      }
  }
}